change the logic
