+++ package.json
@@ -10,7 +10,8 @@
   "devDependencies": {
     "@faker-js/faker": "^9.9.0",
     "@playwright/test": "^1.54.2",
-    "@types/node": "^24.3.0"
+    "@types/node": "^24.3.0",
+    "tsconfig-paths": "^4.2.0"
   },
   "dependencies": {
     "dotenv": "^17.2.1"
