+++ pages/projmanagement/customer.page.ts
@@ -1,0 +1,16 @@
+import { expect, Page } from '@playwright/test';
+
+export class CustomerPage {
+  constructor(private readonly page: Page) {}
+
+  // Customers
+  async createCustomer(customerName: string, description: string) {
+    await this.page.getByRole('button', { name: 'add_circle_outline Create' }).click();
+    await this.page.getByRole('textbox', { name: 'Customer Name' }).click();
+    await this.page.getByRole('textbox', { name: 'Customer Name' }).fill(customerName);
+    await this.page.getByRole('textbox', { name: 'Description' }).click();
+    await this.page.getByRole('textbox', { name: 'Description' }).fill(description);
+    await this.page.getByTestId('stBaseButton-primary').click();
+  }
+
+}
