+++ pages/projmanagement/index.ts
@@ -1,0 +1,3 @@
+export * from './customer.page';
+export * from './project.page';
+export * from './process.page';
