+++ tests/tobe-delete.spec.ts
@@ -1,7 +1,8 @@
+/*
 import { test } from '@playwright/test';
 import { faker } from '@faker-js/faker';
 import { LoginPage } from '../pages/login.page';
-import { ManagementPage } from '../pages/projectmanagement.page';
+import { CustomerPage, Projectpage, ProcessPage } from '../pages'; 
 // import 'dotenv/config';
 
 test('login + create customer/project/process flow', async ({ page }) => {
@@ -17,7 +18,9 @@ test('login + create customer/project/process flow', async ({ page }) => {
   const processName = `${faker.word.noun()}-${faker.number.int(999)}`;
 
   const login = new LoginPage(page);
-  const app = new ManagementPage(page);
+  const customerPage = new CustomerPage(page);
+  const projectPage = new Projectpage(page);
+  const processPage = new ProcessPage(page);
 
   await login.open();                   // uses baseURL from config
   await login.signIn(email, password);  // pass env creds
@@ -32,3 +35,4 @@ test('login + create customer/project/process flow', async ({ page }) => {
   await app.expectSidebarHas(projectName, processName);
   await app.openFrameAssignmentAndExpect();
 });
+*/
