+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "CommonJS",
+    "baseUrl": ".",
+    "paths": {
+      "@pages": ["pages/index.ts"],
+      "@pages/*": ["pages/*"]
+    }
+  },
+  "include": ["pages", "tests", "playwright.config.ts"]
+}
